Change previewCarousel to scroll using the scrollLeft property
scrollIntoView includes vertical scrolling on top of horizontal
scrolling, and is inconsistent in completing the horizontal scrolling as
a result. So change to pure horizontal scrolling by calculating and
setting the scrollLeft property.

To do this cleanly, we added use of jquery to get at the properties, and
use refCallbacks to get references to the game previews